backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==3.2.2
beautifulsoup4==4.13.4
billiard==4.2.1
bs4==0.0.2
celery==5.5.3
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click>=8.0.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.0
filelock==3.18.0
frozenlist==1.7.0
graphql-core==3.2.6
h11==0.16.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
kombu==5.5.4
lxml==6.0.0
multidict==6.6.3
numpy>=1.20.0
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
rsa==4.9.1
scipy>=1.7.0
sgqlc==16.5
shinny-structlog==0.0.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
tqdm==4.67.1
tqsdk==3.8.3
tqsdk-ctpse==1.0.2
tqsdk-sm==1.0.5
tushare==1.4.21
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
psycopg2-binary==2.9.9